requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
jinja2==3.1.4
markdown==3.7
pydantic-settings==2.6.1
aiofiles==24.1.0
python-dotenv==1.0.1
webauthn==2.7.1
